chore: move CheckIfTokenUserIdIsEqualToProvidedService to its own file
