+++ src/domain/services/index.js
@@ -1,5 +1,7 @@
+const CheckIfTokenUserIdIsEqualToProvidedService = require('./check-if-token-user-id-is-equal-to-provided-service');
 const LoginService = require('./login-service');
 
 module.exports = {
+  CheckIfTokenUserIdIsEqualToProvidedService,
   LoginService
 };
